viton-service/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Pillow>=8.3.0
numpy>=1.21.0
scipy>=1.7.0
flask>=2.0.0
flask-cors>=3.0.10
# or
# fastapi>=0.68.0
# uvicorn>=0.15.0
scikit-image>=0.18.0